requirements.txt
Flask
faiss-cpu
numpy
pandas
scikit-learn
SQLAlchemy
